# src/Server/IoTGateway/modules/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.5",
    "title": "REST API Test Docs",
    "description": "該文件可提供在該專案中所需提供的功能"
  },
  "host": "localhost:3095",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/read/UserCustomValueStatus": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "type": "string"
          },
          {
            "name": "ValueName",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/read/UserCustomValueRec": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/set/UserCustomValue": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "ValueName": {
                  "example": "any"
                },
                "num": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}